Update from long? to decimal? for previousVolume (#114)

+++ IEXSharp/Model/Shared/Response/Quote.cs
@@ -35,7 +35,7 @@ namespace IEXSharp.Model.Shared.Response
 		public decimal? extendedChangePercent { get; set; }
 		public long? extendedPriceTime { get; set; }
 		public decimal? previousClose { get; set; }
-		public long? previousVolume { get; set; }
+		public decimal? previousVolume { get; set; }
 		public decimal? change { get; set; }
 		public decimal? changePercent { get; set; }
 		public decimal? volume { get; set; }
